% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions-Params.R
\name{ExtractIndividualCellParam}
\alias{ExtractIndividualCellParam}
\title{#' @export
ExtractCellPeaksParam <- function(
    mz_tol = 20,
    mobility_tol = 0.05,
    mobility_tol_ref = 0.01,
    mobility_tol_int = 0.01,
    smooth_window = 10,
    peak_span_eim = 11,
    min_points = 5,
    snthreshold = 3,
    n_skip = 1,
    min_intensity = 0,
    skip_invalid_peaks = TRUE,
    rerun = FALSE
) {
  new("ExtractCellPeaksParam",
      mz_tol = mz_tol,
      mobility_tol = mobility_tol,
      mobility_tol_ref = mobility_tol_ref,
      mobility_tol_int = mobility_tol_int,
      smooth_window = smooth_window,
      peak_span_eim = peak_span_eim,
      min_points = min_points,
      snthreshold = snthreshold,
      n_skip = n_skip,
      min_intensity = min_intensity,
      skip_invalid_peaks = skip_invalid_peaks,
      rerun = rerun)
}}
\usage{
ExtractIndividualCellParam(
  mz_tol_ppm = 20,
  mobility_tol = 0.04,
  minimal_cell_number = 10,
  pool_size = 100,
  rerun = FALSE
)
}
\description{
#' @export
ExtractCellPeaksParam <- function(
    mz_tol = 20,
    mobility_tol = 0.05,
    mobility_tol_ref = 0.01,
    mobility_tol_int = 0.01,
    smooth_window = 10,
    peak_span_eim = 11,
    min_points = 5,
    snthreshold = 3,
    n_skip = 1,
    min_intensity = 0,
    skip_invalid_peaks = TRUE,
    rerun = FALSE
) {
  new("ExtractCellPeaksParam",
      mz_tol = mz_tol,
      mobility_tol = mobility_tol,
      mobility_tol_ref = mobility_tol_ref,
      mobility_tol_int = mobility_tol_int,
      smooth_window = smooth_window,
      peak_span_eim = peak_span_eim,
      min_points = min_points,
      snthreshold = snthreshold,
      n_skip = n_skip,
      min_intensity = min_intensity,
      skip_invalid_peaks = skip_invalid_peaks,
      rerun = rerun)
}
}
