% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions-Params.R
\name{ExtractSCIMDataParam}
\alias{ExtractSCIMDataParam}
\title{parameters for extracting EIM in single-cell events}
\usage{
ExtractSCIMDataParam(
  mz_tol = 20,
  mobility_tol = 0.05,
  min_intensity = 0,
  min_points = 10,
  n_skip = 1,
  smooth_window = 10,
  peak_span_eim = 11,
  snthreshold = 3,
  skip_invalid_peaks = TRUE,
  rerun = FALSE
)
}
\description{
parameters for extracting EIM in single-cell events
}
