\name{MetCell-package}
\alias{MetCell-package}
\alias{MetCell}
\docType{package}
\title{
\packageTitle{MetCell}
}
\description{
\packageDescription{MetCell}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{MetCell}
\packageIndices{MetCell}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{MetCell}

Maintainer: \packageMaintainer{MetCell}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
# simple examples of the most important functions
}
