% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{_PACKAGE}
\name{MetCell}
\alias{MetCell}
\title{Processing MS data}
\description{
This package processes MS data for peak detection and MSMS spectra assignment
to generate feature table and corresponding MSMS spectra.
}
\author{
Mingdu Luo(luomd@sioc.ac.cn), Yandong Yin (\email{yinyandong@sioc.ac.cn})
}
