% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions-Params.R
\name{QueryTimsDataSegmentParam}
\alias{QueryTimsDataSegmentParam}
\title{parameters for querying TIMS data}
\usage{
QueryTimsDataSegmentParam(
  opentims_thread = 2,
  time_limit_table = "time_limit_table.csv",
  pool_size = 2000L,
  overlaps = 2L,
  rerun = FALSE
)
}
\description{
parameters for querying TIMS data
}
