% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions-Params.R
\name{TrimSCTimsDataParam}
\alias{TrimSCTimsDataParam}
\title{Trim data according to the provided time limit table}
\usage{
TrimSCTimsDataParam(time_limit_table = "time_limit_table.csv", rerun = FALSE)
}
\arguments{
\item{time_limit_table}{\code{character} path to the time limit table}

\item{rerun}{\code{logical} if rerun the process}
}
\description{
Trim data according to the provided time limit table
}
