% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions-Params.R
\name{DiscoverSCEventsParam}
\alias{DiscoverSCEventsParam}
\title{Find single cell event according to the provided single cell marker}
\usage{
DiscoverSCEventsParam(
  marker_mz = 760.5851,
  marker_mobility = 1.355,
  mz_tolerance_ppm = 20,
  mobility_range = 0.015,
  intensity_abs_threshold_upper = NULL,
  intensity_abs_threshold_lower = NULL,
  marker_eic_peak_span = 5
)
}
\arguments{
\item{marker_mz}{\code{numeric} m/z of the single-cell event marker}

\item{intensity_abs_threshold_upper}{\code{numeric} relative intensity upper threshold of the single-cell event marker}

\item{intensity_abs_threshold_lower}{\code{numeric} relative intensity lower threshold of the single-cell event marker}

\item{marker_ccs}{\code{numeric} ccs of the single-cell event marker}
}
\description{
Find single cell event according to the provided single cell marker
}
